Готово, вроде как

+++ ProjectControl/Controllers/TasksController.cs
@@ -18,19 +18,24 @@ namespace ProjectControl.Controllers
             return UserControll.LoggedAs == null || !UserControll.LoggedAs.IsAdmin;
         }
 
-        public TaskResponce Get()
+        public TaskResponce Get(bool getAll = true)
         {
             TaskResponce responce = new TaskResponce();
             responce.StatusCode = 200;
 
-            if (IsNotAccessable())
+            User user = UserControll.LoggedAs;
+
+            if (user == null)
             {
                 responce.StatusCode = 403;
                 responce.ErrorMessage.Add("No access rights");
             }
             else
             {
-                responce.Tasks.AddRange(db.Tasks);
+                if (user.IsAdmin && getAll)
+                    responce.Tasks.AddRange(db.Tasks);
+                else
+                    responce.Tasks.AddRange(db.Tasks.Where(x => x.CreatorLogin == user.Login));
             }
 
             return responce;
@@ -39,27 +44,38 @@ namespace ProjectControl.Controllers
         public TaskResponce Get(int id)
         {
             TaskResponce responce = new TaskResponce();
+            User user = UserControll.LoggedAs;
             responce.StatusCode = 200;
 
-            if (IsNotAccessable())
+            if (user == null)
             {
                 responce.StatusCode = 403;
                 responce.ErrorMessage.Add("No access rights");
             }
             else
             {
-                responce.Tasks.AddRange(db.Tasks.Where(x => x.Id == id));
+                Task task = db.Tasks.Where(x => x.Id == id).FirstOrDefault();
+                if (task.CreatorLogin != user.Login && !user.IsAdmin)
+                {
+                    responce.StatusCode = 403;
+                    responce.ErrorMessage.Add("No access rights");
+
+                    return responce;
+                }
+
+                responce.Tasks.Add(task);
             }
 
             return responce;
         }
 
-        public TaskResponce Get(string title, string creator, string status, int projectId = -1)
+        public TaskResponce Get(string title, string status, int projectId = -1, string creator = "", bool getAll = true)
         {
             TaskResponce responce = new TaskResponce();
+            User user = UserControll.LoggedAs;
             responce.StatusCode = 200;
 
-            if (IsNotAccessable())
+            if (user == null)
             {
                 responce.StatusCode = 403;
                 responce.ErrorMessage.Add("No access rights");
@@ -74,21 +90,32 @@ namespace ProjectControl.Controllers
                     status = "";
 
                 if (projectId == -1)
-                    responce.Tasks.AddRange(db.Tasks.ToList().Where(x => x.Name.Contains(title) && x.CreatorLogin.Contains(creator) && x.Status.Contains(status)));
+                {
+                    if (user.IsAdmin && getAll)
+                        responce.Tasks.AddRange(db.Tasks.ToList().Where(x => x.Name.Contains(title) && x.CreatorLogin.Contains(creator) && x.Status.Contains(status)));
+                    else
+                        responce.Tasks.AddRange(db.Tasks.ToList().Where(x => x.Name.Contains(title) && x.Status.Contains(status) && x.CreatorLogin == user.Login));
+                }
                 else
-                    responce.Tasks.AddRange(db.Tasks.ToList().Where(x => x.Name.Contains(title) && x.CreatorLogin.Contains(creator) && x.Status.Contains(status) && x.ProjectId == projectId));
+                {
+                    if (user.IsAdmin && getAll)
+                        responce.Tasks.AddRange(db.Tasks.ToList().Where(x => x.Name.Contains(title) && x.CreatorLogin.Contains(creator) && x.Status.Contains(status) && x.ProjectId == projectId));
+                    else
+                        responce.Tasks.AddRange(db.Tasks.ToList().Where(x => x.Name.Contains(title) && x.Status.Contains(status) && x.ProjectId == projectId && x.CreatorLogin == user.Login));
+                }
             }
 
             return responce;
         }
 
-        public TaskResponce Post(Task _task)
+        public TaskResponce Post(Task _task, bool autoStatus = true)
         {
             TaskResponce responce = new TaskResponce();
             Project _project = db.Projects.Where(x => x.Id == _task.ProjectId).FirstOrDefault();
+            User user = UserControll.LoggedAs;
             responce.StatusCode = 200;
 
-            if (IsNotAccessable())
+            if (user == null)
             {
                 responce.StatusCode = 403;
                 responce.ErrorMessage.Add("No access rights");
@@ -130,6 +157,8 @@ namespace ProjectControl.Controllers
                 {
                     _task.CreatorLogin = UserControll.LoggedAs.Login;
                     _task.CreatedTime = DateTime.Now;
+                    if (autoStatus)
+                        _task.IsAccepted = null;
 
                     db.Tasks.Add(_task);
                     db.SaveChanges();
@@ -143,16 +172,25 @@ namespace ProjectControl.Controllers
         {
             TaskResponce responce = new TaskResponce();
             Task task = db.Tasks.Where(x => x.Id == _task.Id).FirstOrDefault();
+            User user = UserControll.LoggedAs;
             Project _project = db.Projects.Where(x => x.Id == task.ProjectId).FirstOrDefault();
             responce.StatusCode = 200;
 
-            if (IsNotAccessable())
+            if (user == null)
             {
                 responce.StatusCode = 403;
                 responce.ErrorMessage.Add("No access rights");
             }
             else
             {
+                if (user.Login != task.CreatorLogin && !user.IsAdmin)
+                {
+                    responce.StatusCode = 403;
+                    responce.ErrorMessage.Add("No access rights");
+
+                    return responce;
+                }
+
                 if (string.IsNullOrWhiteSpace(_task.Name))
                 {
                     responce.StatusCode = 400;
@@ -179,6 +217,11 @@ namespace ProjectControl.Controllers
                     responce.StatusCode = 400;
                     responce.ErrorMessage.Add("Task time is not in bounds of project time");
                 }
+                if (_task.StartTime > _task.EndTime)
+                {
+                    responce.StatusCode = 400;
+                    responce.ErrorMessage.Add("End time is less than start time");
+                }
 
                 if (responce.StatusCode == 200)
                 {
@@ -199,9 +242,10 @@ namespace ProjectControl.Controllers
         public TaskResponce Delete(int id)
         {
             TaskResponce responce = new TaskResponce();
+            User user = UserControll.LoggedAs;
             responce.StatusCode = 200;
 
-            if (IsNotAccessable())
+            if (user == null)
             {
                 responce.StatusCode = 403;
                 responce.ErrorMessage.Add("No access rights");
@@ -211,6 +255,14 @@ namespace ProjectControl.Controllers
                 Task task = db.Tasks.Where(x => x.Id == id).FirstOrDefault();
                 if (task == null)
                     return responce;
+
+                if (user.Login != task.CreatorLogin && !user.IsAdmin)
+                {
+                    responce.StatusCode = 403;
+                    responce.ErrorMessage.Add("No access rights");
+
+                    return responce;
+                }
 
                 db.Tasks.Remove(task);
                 db.SaveChanges();
